[auto] Separar aprobación de negocios
