+++ app/composeApp/src/commonMain/kotlin/DIManager.kt
@@ -38,6 +38,7 @@ public const val CHANGE_PASSWORD = "changePassword"
 public const val PASSWORD_RECOVERY = "passwordRecovery"
 public const val CONFIRM_PASSWORD_RECOVERY = "confirmPasswordRecovery"
 public const val REGISTER_BUSINESS = "registerBusiness"
+public const val REVIEW_BUSINESS = "reviewBusiness"
 
 const val LOGIN_PATH = "/login"
 
@@ -65,6 +66,7 @@ class DIManager {
                 bindSingleton(tag = PASSWORD_RECOVERY) { PasswordRecoveryScreen() }
                 bindSingleton(tag = CONFIRM_PASSWORD_RECOVERY) { ConfirmPasswordRecoveryScreen() }
                 bindSingleton(tag = REGISTER_BUSINESS) { RegisterBusinessScreen() }
+                bindSingleton(tag = REVIEW_BUSINESS) { ReviewBusinessScreen() }
 
                 bindSingleton (tag = SCREENS) {
                     arrayListOf<Screen>(
@@ -79,7 +81,8 @@ class DIManager {
                         instance(tag = CHANGE_PASSWORD),
                         instance(tag = PASSWORD_RECOVERY),
                         instance(tag = CONFIRM_PASSWORD_RECOVERY),
-                        instance(tag = REGISTER_BUSINESS)
+                        instance(tag = REGISTER_BUSINESS),
+                        instance(tag = REVIEW_BUSINESS)
                     )
                 }
