+++ app/composeApp/src/commonMain/kotlin/ui/sc/RegisterBusinessViewModel.kt
@@ -2,8 +2,6 @@ package ui.sc
 
 import DIManager
 import asdo.ToDoRegisterBusiness
-import asdo.ToDoReviewBusinessRegistration
-import asdo.ToGetBusinesses
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
@@ -16,18 +14,14 @@ import org.kodein.log.newLogger
 class RegisterBusinessViewModel : ViewModel() {
     private val logger = LoggerFactory.default.newLogger<RegisterBusinessViewModel>()
     private val register: ToDoRegisterBusiness by DIManager.di.instance()
-    private val review: ToDoReviewBusinessRegistration by DIManager.di.instance()
-    private val getBusinesses: ToGetBusinesses by DIManager.di.instance()
 
     var state by mutableStateOf(UIState())
     var loading by mutableStateOf(false)
-    var pending by mutableStateOf(listOf<String>())
 
     data class UIState(
         val name: String = "",
         val email: String = "",
-        val description: String = "",
-        val twoFactorCode: String = ""
+        val description: String = ""
     )
 
     override fun getState(): Any = state
@@ -45,8 +39,7 @@ class RegisterBusinessViewModel : ViewModel() {
         inputsStates = mutableMapOf(
             entry(UIState::name.name),
             entry(UIState::email.name),
-            entry(UIState::description.name),
-            entry(UIState::twoFactorCode.name)
+            entry(UIState::description.name)
         )
     }
 
@@ -54,24 +47,4 @@ class RegisterBusinessViewModel : ViewModel() {
         register.execute(state.name, state.email, state.description)
             .onSuccess { logger.info { "Negocio registrado: ${'$'}{state.name}" } }
             .onFailure { error -> logger.error { "Error registrando negocio: ${'$'}{error.message}" } }
-
-    suspend fun approve(business: String) =
-        review.execute(business, "approved", state.twoFactorCode)
-            .onSuccess { logger.info { "Negocio aprobado: ${'$'}business" } }
-            .onFailure { error -> logger.error { "Error aprobando ${'$'}business: ${'$'}{error.message}" } }
-
-    suspend fun reject(business: String) =
-        review.execute(business, "rejected", state.twoFactorCode)
-            .onSuccess { logger.warning { "Negocio rechazado: ${'$'}business" } }
-            .onFailure { error -> logger.error { "Error rechazando ${'$'}business: ${'$'}{error.message}" } }
-
-    suspend fun loadPending() {
-        logger.debug { "Cargando negocios pendientes" }
-        getBusinesses.execute("PENDING")
-            .onSuccess {
-                pending = it.businesses
-                logger.info { "Pendientes obtenidos: ${'$'}{pending.size}" }
-            }
-            .onFailure { error -> logger.error { "Error cargando pendientes: ${'$'}{error.message}" } }
-    }
 }
